Adds a mixin that Fixes JER biome lookup for modded biomes
Addresses an issue where Just Enough Resources (JER) fails to correctly identify and display modded biomes.

JER was using a vanilla-only registry, excluding modded biomes from its biome list. This commit introduces a mixin that replaces
the vanilla registry lookup with the current level's registry, which includes all biomes.

the mixin is conditionally applied only when JER is present. Fallback methods
are provided to use the vanilla registry if the level is not available.

This is an issue thats been reported for over a year in JER and never fixed.